% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/BC.R
\name{BC}
\alias{BC}
\title{Bayesian Combination}
\usage{
BC(sat, gauge, longlat = TRUE, cross.val = FALSE)
}
\arguments{
\item{sat}{is a list
sat[[2]] is a data.frame with dimensions nrow=no of satellite
pixels, containing columns x=x-coordinates, y=y-coordinates
sat  [[1]] is a zoo object with dimensions ncol=no of satellite
pixels, nrow=no of timestep}

\item{gauge}{is a list:
gauge[[2]] is a data.frame with dimensions nrow=no of stations,
containing columns x=x-coordinates, y=y-coordinates
gauge[[1]] is a zoo object with dimensions ncol=no of stations,
nrow=no of timestep}

\item{longlat}{is a flag to describe the coordinate grids of spatial data.
Defaults to TRUE i.e. on long-lat notation}

\item{cross.val}{option TRUE=in cross validation mode; default FALSE}
}
\value{
Zs merging output - zoo object with dimensions == dimensions sat
[["ts""]

crossval cross validation- zoo object with dimensions == dimensions
gauge[["ts""]
}
\description{
This function merges satellite and Ordinary-Kriged rain
gauge fields using Bayesian inference
References: Todini, E. (2001). A Bayesian technique for
conditioning radar precipitation estimates to rain-gauge measurements.
Hydrology and Earth System Sciences, 5(2):187-199.
PROGEA Srl (2009). RAINMUSIC, User manual & references. PROGEA Srl, Bologna.

#'
}
\examples{
# BC_out <- BC(sat, gauge)
# BC_cv  <- BC(sat, gauge, cross.val=TRUE)
}
\author{
Zed Zulkafli, Daniele Nerini
}

