% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/colocate_gauge.R
\name{colocate.gauge}
\alias{colocate.gauge}
\title{Colocate gauges}
\usage{
colocate.gauge(gauge, coords, longlat = TRUE)
}
\arguments{
\item{gauge}{is a list:
gauge[[2]] is a data.frame with dimensions nrow=no of stations,
containing columns x=x-coordinates, y=y-coordinates
gauge[[1]] is a zoo object with dimensions ncol=no of stations,
nrow=no of timestep}

\item{coords}{is sat  [[2]] is a data.frame with dimensions nrow=no of
satellite pixels, containing columns x=x-coordinates, y=y-coordinates}

\item{longlat}{is a flag for geospatial projection}
}
\value{
gauge see param description
}
\description{
This function colocates and average rain gauges estimates to the
nearest center cell of a target interpolation grid
}
\examples{
# gauge <- colocage.gauge(gauge, sat[[2]], longlat=TRUE)
}
\author{
Zed Zulkafli
}

