% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/crossval_score.R
\name{crossval_score}
\alias{crossval_score}
\title{Cross validation scores}
\usage{
crossval_score(results_CV, gauge)
}
\arguments{
\item{results_CV}{is a list of the output from one or multiple merging functions run in cross-validation mode}

\item{gauge}{is a list:
gauge[[2]] is a data.frame with dimensions nrow=no of stations,
containing columns x=x-coordinates, y=y-coordinates
gauge[[1]] is a zoo object with dimensions ncol=no of stations,
nrow=no of timestep}
}
\value{
scores array containing performance indices for the cross-validation
}
\description{
This function calculates performance indices for
for the cross-validation output for comparison between the methods
}
\examples{
# results <- crossval_score(results_CV,gauge)
}
\author{
Zed Zulkafli
}

