% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/KED.R
\name{KED}
\alias{KED}
\title{Kriging with External Drift}
\usage{
KED(sat, gauge, longlat = TRUE, cross.val = FALSE)
}
\arguments{
\item{sat}{is a list
sat[[2]] is a data.frame with dimensions nrow=no of satellite
pixels, containing columns x=x-coordinates, y=y-coordinates
sat  [[1]] is a zoo object with dimensions ncol=no of satellite
pixels, nrow=no of timestep}

\item{gauge}{is a list:
gauge[[2]] is a data.frame with dimensions nrow=no of stations,
containing columns x=x-coordinates, y=y-coordinates
gauge[[1]] is a zoo object with dimensions ncol=no of stations,
nrow=no of timestep}

\item{longlat}{is a flag to describe the coordinate grids of spatial data.
Defaults to TRUE i.e. on long-lat notation}

\item{cross.val}{option TRUE=in cross validation mode; default FALSE}
}
\value{
Zs merging output - zoo object with dimensions == dimensions sat
[["ts""]

crossval cross validation- zoo object with dimensions == dimensions gauge
[["ts""]
}
\description{
This function performs a Kriging interpolation of rain gauge
estimates using satellite estimates as an external drift
}
\examples{
# KED_out <- KED(sat, gauge)
# KED_cv  <- KED(sat, gauge, cross.val=TRUE)
}
\author{
Zed Zulkafli, Bastian Manz
}

