% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NN_interp.R
\name{NN.interp}
\alias{NN.interp}
\title{Nearest-neighbour interpolator}
\usage{
NN.interp(sat, coords)
}
\arguments{
\item{sat}{is a list
sat[[2]] is a data.frame with dimensions nrow=no of satellite
pixels, containing columns x=x-coordinates, y=y-coordinates, ID=pixel ID
sat  [[1]] is a zoo object with dimensions ncol=no of satellite
pixels, nrow=no of timestep}

\item{coords}{is a data.frame with dimensions nrow=no of target grid pixels,
containing columns x=x-coordinates, y=y-coordinates}
}
\value{
sat see param description
}
\description{
This function interpolates satellite data to a new grid using
the nearest-neighbour interpolation
}
\examples{
# sat <- NN.interp(sat, coords)
}
\author{
Zed Zulkafli
}

