% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/OK.R
\name{OK}
\alias{OK}
\title{Ordinary Kriging}
\usage{
OK(coords, gauge, cross.val = FALSE)
}
\arguments{
\item{coords}{is a data.frame with dimensions nrow=no of target grid pixels,
containing columns x=x-coordinates, y=y-coordinates}

\item{gauge}{is a list:
gauge[[2]] is a data.frame with dimensions nrow=no of stations,
containing columns x=x-coordinates, y=y-coordinates
gauge[[1]] is a zoo object with dimensions ncol=no of stations,
nrow=no of timestep}

\item{cross.val}{option TRUE=in cross validation mode; default FALSE}
}
\value{
Zs merging output - zoo object with dimensions == dimensions sat
[["ts""]

crossval cross validation- zoo object with dimensions == dimensions
gauge[["ts""]
}
\description{
This function interpolate rain gauge data into the target grid
using Ordinary Kriging at every time step
}
\examples{
# OK_out <- OK(sat, gauge)
# OK_cv  <- OK(sat, gauge, cross.val=TRUE)
}
\author{
Zed Zulkafli
}

